requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.4.0
click>=8.1.6
tqdm>=4.66.1

# Git interaction
GitPython>=3.1.40

# LLM API clients
llama-cpp-python>=0.2.0
anthropic>=0.5.0  # As an alternative option

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# UI components
rich>=13.4.2
streamlit>=1.30.0  # For web interface

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
